Advance bit_sample_timer by half an interrupt time
The leading edge of the start bit occurs, on average, half of an
interrupt time before we detect it. This advances our sampling timer by
that amount of time to ensure that we sample bits closer to their
centers.

This doesn't make much of a difference except when the int_rate/bit_rate
ratio is low (say, 3 or less).