Add a state to wait for stop bit
If we detect a framing error (the stop bit is 0, while it should be 1),
add the received byte to the queue once and then move to a new state to
wait for a real stop bit.

A break condition (in which the input pin is pulled low for longer than
one UART frame) will cause SoftRxUart to add a single 0x00 byte to the
queue and then wait until the pin goes high again before it resumes
waiting for the next start bit. Before this change, SoftRxUart would
have added 0x00 bytes to the queue for as long as the break condition
persisted.

Besides preventing a break condition from filling up the receive queue,
this also somewhat improves resynchronizing after data corruption, as
SoftRxUart will not look for a start bit until it receives a stop bit.